move grpc server to separate package
